fix: fixup, put GasSettingDialog under BottomDrawer
